+++ packages/browser-wallet/src/popup/pages/Allowlist/Allowlist.scss
@@ -1,20 +1,12 @@
-.empty-allowlist-settings-page {
+.allowlist-page,
+.allowlist-empty {
     height: 100%;
     width: 100%;
     background-color: $color-bg;
     padding: 0 rem(10px);
-    display: flex;
-    align-items: center;
-    text-align: center;
-    color: $color-text-faded;
 }
 
-.allowlist-settings-page {
-    height: 100%;
-    width: 100%;
-    background-color: $color-bg;
-    padding: 0 rem(10px);
-
+.allowlist-page {
     &__link-icon {
         width: rem(20px);
         height: rem(10px);
@@ -29,56 +21,79 @@
     }
 }
 
-.connect-accounts-request-accounts {
-    padding-left: rem(10px);
-    padding-right: rem(10px);
-    width: 100%;
-    min-height: rem(175px);
-    height: rem(175px);
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: rem(15px);
-    background: $color-input-bg;
-    border: rem(1px) solid #dae4ee;
-    border-top: none;
-    border-radius: 0 0 rem(10px) rem(10px);
-    scrollbar-width: thin;
-    overflow-y: auto;
+.allowlist-empty {
+    display: flex;
+    align-items: center;
+    text-align: center;
+    color: $color-text-faded;
+}
 
-    &__account-item {
-        padding-top: rem(5px);
-        padding-bottom: rem(5px);
-        font-weight: $font-weight-light;
+.allowlist-entry-view {
+    &__header {
+        width: 100%;
+        margin-top: rem(8px);
+        padding: rem(10px);
+        padding-top: rem(0);
+        background: #eaf0f5;
+        border: rem(1px) solid #dae4ee;
+        border-radius: rem(10px) rem(10px) 0 0;
+    }
 
-        &:not(:last-child) {
-            border-bottom: rem(1px) dashed #dae4ee;
-        }
+    &__mode-description {
+        margin-top: rem(10px);
+    }
 
-        &__primary {
-            position: relative;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
+    &__accounts {
+        width: 100%;
+        min-height: rem(160px);
+        height: rem(160px);
+        padding-left: rem(10px);
+        padding-right: rem(10px);
+        margin-left: auto;
+        margin-right: auto;
+        margin-bottom: rem(15px);
+        background: $color-input-bg;
+        border: rem(1px) solid #dae4ee;
+        border-top: none;
+        border-radius: 0 0 rem(10px) rem(10px);
+        scrollbar-width: thin;
+        overflow-y: auto;
 
-        &__check-box {
-            position: absolute;
-            right: rem(0);
-            bottom: rem(-12.5px);
-            width: rem(15px);
+        &__item {
+            padding-top: rem(5px);
+            padding-bottom: rem(5px);
+            font-weight: $font-weight-light;
 
-            input {
-                margin: 0;
+            &:not(:last-child) {
+                border-bottom: rem(1px) dashed #dae4ee;
             }
-        }
 
-        &__secondary {
-            font-size: rem(8px);
+            &__primary {
+                position: relative;
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+            }
+
+            &__check-box {
+                position: absolute;
+                right: rem(0);
+                bottom: rem(-12.5px);
+                width: rem(15px);
+
+                input {
+                    margin: 0;
+                }
+            }
+
+            &__secondary {
+                font-size: rem(8px);
+            }
         }
     }
 }
 
-.allow-list-edit-page {
+.allow-list-editor {
     height: 100%;
     padding: rem(10px);
     background-color: $color-bg;
