+++ packages/browser-wallet/src/popup/pages/Allowlist/i18n/en.ts
@@ -1,13 +1,16 @@
 const t = {
     empty: 'You do not currently have any services in your allowlist.',
-    editor: {
+    entry: {
         header: 'Allowlisting a service',
         description:
             'Allowlisting a service means that it can request identity proofs and signatures from selected accounts.',
+        addDescription: 'Select accounts to share with the service below.',
+        modifyDescription:
+            'You can modify which accounts are accessible to the service below. Any allowlisted service can request proofs.',
     },
     view: {
-        remove: 'Remove service from allowlist',
-        update: 'Update allowlist',
+        remove: 'Remove',
+        update: 'Update',
     },
 };
