--- packages/browser-wallet/src/popup/pages/ConnectAccountsRequest/ConnectAccountsRequest.scss
@@ -1,9 +1,0 @@
-.connect-accounts-request {
-    width: 100%;
-    margin-top: rem(8px);
-    padding: rem(10px);
-    padding-top: rem(0);
-    background: #eaf0f5;
-    border: rem(1px) solid #dae4ee;
-    border-radius: rem(10px) rem(10px) 0 0;
-}
