+++ packages/browser-wallet/src/popup/styles/_components.scss
@@ -27,7 +27,6 @@
 @import '../pages/SignMessage/SignMessage';
 @import '../pages/Setup/Setup';
 @import '../pages/ConnectionRequest/ConnectionRequest';
-@import '../pages/ConnectAccountsRequest/ConnectAccountsRequest';
 @import '../pages/Settings/Settings';
 @import '../pages/Identity/Identity';
 @import '../pages/IdentityIssuance/IdentityIssuance';
